requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2023.11.17
cloudinary==1.37.0
Django==5.0.1
django-cors-headers==4.3.1
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
pyotp==2.9.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rpds-py==0.16.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
